Clock_v1
This is a more developed version of the alarm clock. I have a deeper understanding of how the mills() function does work. I have implemented  a light fade system that fade in and out, by using a sine wave that ranges from 0 - 1; and maps that function to a specific color range, anywhere between 0-255, using a custom function that uses a float instead of a long/int; mapf, that is the same to the arduino api, since it cant take in floats, sad. Excited I also created a LFO, that I will not use........... since I will remove the speaker from the project. there are some issues where the speaker is affected by the RGB lights fade function. Also implemented a alarm function, a preAlarm function, and created and array that symbolizes hours-minutes-seconds, as argument passed. It all sounds confusing and probably is..........